refatoraçao da store e alguns componentes, adiçao de pontos e rodadas na tabela(real). basico do placar funcionando
