--- store/modules/pontos/index.js
@@ -1,9 +1,0 @@
-export default {
-  namespaced: true,
-  state: {
-
-  },
-  mutations: {},
-  actions: {},
-  getters: {}
-}
